backend/requirements.txt
# ─── Django core ──────────────────────────────────────────────────────────────
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# ─── Celery + Redis (tâches planifiées) ───────────────────────────────────────
celery>=5.3
redis>=5.0

# ─── HTTP (appels SFMC) ───────────────────────────────────────────────────────
requests>=2.31

# ─── Config (.env) ────────────────────────────────────────────────────────────
python-decouple>=3.8

# ─── Serveur production ───────────────────────────────────────────────────────
gunicorn>=21.2

# ─── Base de données production (PostgreSQL) ──────────────────────────────────
psycopg2-binary>=2.9
